Removed buggy LRE compression from 3D raster library as discussed here: http://trac.osgeo.org/grass/ticket/1752 3D raster modules have been adjusted to the API changes. Tests have been improved.
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@55462 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/raster3d/tileread.c
@@ -88,7 +88,7 @@ Rast3d_xdrTile2tile(RASTER3D_Map * map, void *tile, int rows, int cols, int dept
 
 static int Rast3d_readTileUncompressed(RASTER3D_Map * map, int tileIndex, int nofNum)
 {
-    int nofBytes;
+    size_t nofBytes;
 
     nofBytes = nofNum * map->numLengthExtern;
     nofBytes = RASTER3D_MIN(nofBytes, map->fileEndPtr - map->index[tileIndex]);
@@ -139,7 +139,7 @@ static int Rast3d_readTileCompressed(RASTER3D_Map * map, int tileIndex, int nofN
  *  \param tile
  *  \param type
  *  \return 1 ... if successful,
- *          0 ... otherwise.
+ *          0 ... otherwise
  */
 
 int Rast3d_read_tile(RASTER3D_Map * map, int tileIndex, void *tile, int type)
